Create an inversion of control container for initializing the CDI container
