+++ api/src/main/java/devsimiyu/samserverless/api/App.java
@@ -7,6 +7,7 @@ import com.amazonaws.services.lambda.runtime.Context;
 import com.amazonaws.services.lambda.runtime.RequestHandler;
 import com.amazonaws.services.lambda.runtime.events.APIGatewayProxyRequestEvent;
 import com.amazonaws.services.lambda.runtime.events.APIGatewayProxyResponseEvent;
+import devsimiyu.samserverless.core.config.IoC;
 import devsimiyu.samserverless.core.model.dto.Address;
 import devsimiyu.samserverless.core.model.dto.User;
 import devsimiyu.samserverless.core.security.Jwt;
@@ -25,15 +26,7 @@ import jakarta.enterprise.inject.spi.Extension;
 public class App implements RequestHandler<APIGatewayProxyRequestEvent, APIGatewayProxyResponseEvent> {
 
     public APIGatewayProxyResponseEvent handleRequest(final APIGatewayProxyRequestEvent input, final Context context) {
-        SeContainerInitializer containerInitializer = SeContainerInitializer.newInstance();
-        containerInitializer.addExtensions(new Extension() {
-            public void addJwt(@Observes AfterBeanDiscovery event) {
-                System.out.println("AfterBeanDiscovery Extension CALLED!!");
-                Jwt token = () -> "jwt primitive token string from main";
-                event.addBean().types(Jwt.class).scope(ApplicationScoped.class).addQualifier(Default.Literal.INSTANCE).createWith(obj -> token);
-            }
-        });
-
+        String token = "jwt primitive token string from main";
         Map<String, String> headers = new HashMap<>();
         headers.put("Content-Type", "application/json");
         headers.put("X-Custom-Header", "application/json");
@@ -41,7 +34,7 @@ public class App implements RequestHandler<APIGatewayProxyRequestEvent, APIGatew
         APIGatewayProxyResponseEvent response = new APIGatewayProxyResponseEvent()
                 .withHeaders(headers);
 
-        try (SeContainer container = containerInitializer.initialize()) {
+        try (SeContainer container = IoC.INSTANCE.initialize(token)) {
             PingService pingService = container.select(PingService.class).get();
             User user = new User();
             Address address = new Address();
@@ -52,15 +45,8 @@ public class App implements RequestHandler<APIGatewayProxyRequestEvent, APIGatew
             address.street = "Gakere";
             String output = String.format("{ \"message\": \"hello world\", \"location\": \"%s\" }", pingService.sayHello(user));
 
-            return response
-                    .withStatusCode(200)
-                    .withBody(output);
+            return response.withStatusCode(200).withBody(output);
 
-        }
-        catch (Exception e) {
-            return response
-                    .withBody("{}")
-                    .withStatusCode(500);
-        }
+        } catch (Exception e) { return response.withBody("{}").withStatusCode(500); }
     }
 }
