+++ core/src/main/java/devsimiyu/samserverless/core/Main.java
@@ -1,5 +1,6 @@
 package devsimiyu.samserverless.core;
 
+import devsimiyu.samserverless.core.config.IoC;
 import devsimiyu.samserverless.core.interceptors.Auth;
 import devsimiyu.samserverless.core.model.dto.Address;
 import devsimiyu.samserverless.core.model.dto.TodoItem;
@@ -20,21 +21,8 @@ import jakarta.enterprise.inject.spi.*;
 public class Main {
 
     public static void main(String[] args) {
-        Extension authSession = new Extension() {
-            public void addJwt(@Observes AfterBeanDiscovery event) {
-                System.out.println("AfterBeanDiscovery Extension CALLED!!");
-                Jwt token = () -> "jwt primitive token string from main";
-                event.addBean().types(Jwt.class).scope(ApplicationScoped.class).addQualifiers(Default.Literal.INSTANCE).createWith(context -> token);
-            }
-        };
-        Extension authInterceptor = new Extension() {
-            public void addJwt(@Observes ProcessAnnotatedType<?> event) {
-                System.out.println("ProcessAnnotatedType Extension CALLED!!");
-                event.configureAnnotatedType().filterMethods(annotatedMethod -> annotatedMethod.isAnnotationPresent(RolesAllowed.class)).forEach(annotatedMethodConfigurator -> annotatedMethodConfigurator.add(Auth.Literal.INSTANCE));
-            }
-        };
-        SeContainerInitializer containerInitializer = SeContainerInitializer.newInstance().addExtensions(authSession, authInterceptor);
-        try (SeContainer container = containerInitializer.initialize()) {
+        String token = "jwt primitive token string from main";
+        try (SeContainer container = IoC.INSTANCE.initialize(token)) {
             PingService pingService = container.select(PingService.class).get();
             System.out.println("Ping Message: " + pingService.ping());
         }
